# Conflicts:
#	whalegoods-erp/whalegoods-erp/src/main/resources/application.yml

+++ whalegoods-erp/whalegoods-erp/src/main/java/com/whalegoods/entity/Device.java
@@ -28,7 +28,7 @@ public class Device extends BaseEntity implements Serializable {
     
     private Byte deviceStatus;
     
-    private Integer upTime;
+    private Long upTime;
     
     private String signCode;
